demo

+++ controllers/paymentControllers.js
@@ -18,20 +18,32 @@ module.exports.ipn = async (req, res) => {
     const payment = new Payment(req.body);
     const tran_id = payment['tran_id'];
     if (payment['status'] === "VALID") {
-        const order = await Order.updateOne({ transaction_id: tran_id }, { status: "Complete" });
-        const cart = await Order.findOne({ transaction_id: tran_id });
-        const items = cart.cartItems;
-        for (let x of items) {
-            const product = await Product.findOne({ _id: x.product });
-            const sold = product.sold + x.count;
-            await Product.updateOne({ _id: x.product }, { sold: sold });
+        // const order = await Order.updateOne({ transaction_id: tran_id }, { status: "Complete" });
+        // const cart = await Order.findOne({ transaction_id: tran_id });
+        // const items = cart.cartItems;
+        // for (let x of items) {
+        //     const product = await Product.findOne({ _id: x.product });
+        //     const sold = product.sold + x.count;
+        //     await Product.updateOne({ _id: x.product }, { sold: sold });
 
-        }
-        await CartItem.deleteMany(order.cartItems);
+        // }
+        // await CartItem.deleteMany(order.cartItems);
 
         //Order Validation API
         axios.get(`https://sandbox.sslcommerz.com/validator/api/validationserverAPI.php?val_id=${payment.val_id}&store_id=${process.env.STORE_ID}&store_passwd=${process.env.STORE_PASSWORD}&format=json`)
-            .then(response => console.log(response.data.status))
+            .then(async (response) => {
+                if (response.data.status === "VALID" || response.data.status === "VALIDATED") {
+                    const order = await Order.updateOne({ transaction_id: tran_id }, { status: "Complete" });
+                    const cart = await Order.findOne({ transaction_id: tran_id });
+                    const items = cart.cartItems;
+                    for (let x of items) {
+                        const product = await Product.findOne({ _id: x.product });
+                        const sold = product.sold + x.count;
+                        await Product.updateOne({ _id: x.product }, { sold: sold });
+                    }
+                    await CartItem.deleteMany(order.cartItems);
+                }
+            })
             .catch(error => console.log(error));
     }
